Move job utils to shared folder
