+++ web_ui/src/features/dataset-import/hooks/use-dataset-import-queries.hook.ts
@@ -20,7 +20,6 @@ import {
     DatasetPrepareForExistingProjectIdentifier,
     DatasetPrepareForExistingProjectResponse,
 } from '../../../core/datasets/dataset.interface';
-import { getIntervalJobHandlers } from '../../../core/datasets/hooks/utils';
 import {
     JobImportDatasetToExistingProjectStatus,
     JobImportDatasetToNewProjectStatus,
@@ -31,6 +30,7 @@ import {
 import { CreateDatasetResponse } from '../../../core/projects/dataset.interface';
 import { NOTIFICATION_TYPE } from '../../../notification/notification-toast/notification-type.enum';
 import { useNotification } from '../../../notification/notification.component';
+import { getIntervalJobHandlers } from '../../../shared/utils';
 import { IntervalJobHandlers } from '../types/dataset-import.interface';
 
 interface StatusJobProps<T> extends IntervalJobHandlers<T> {
